Fixed overflow issues
Discovered a problem: If you didn't read the keypad for a long time, and the user had input more than 15 buttons, there was an overflow error.

It had to do with the incrementing of oldest and newest press variables. The incrementations needed to be ahead of the read/write actions. We also needed some more conditionals to handle the "trailing just one behind scenario".

Also, started to add in the register map stuff.